requirements.txt
Django==1.7.1
PyJWT==0.3.0
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==0.6
django-compressor==1.4
djangorestframework==2.4.4
djangorestframework-jwt==1.0.2
drf-nested-routers==0.9.0
gunicorn==19.1.1
model-mommy==1.2.2
nodeenv==0.11.1
six==1.8.0
static3==0.5.1
